+++ Src/AppLib/VpnHood.AppLib.App/AppConnectionState.cs
@@ -10,6 +10,8 @@ public enum AppConnectionState
     Waiting,
     WaitingForAd,
     Diagnosing,
+    FindingReachableServer,
+    FindingBestServer,
     Connecting,
     Connected,
     Unstable,