+++ Src/AppLib/VpnHood.AppLib.App/Utils/StateHelper.cs
@@ -1,12 +1,13 @@
 ﻿using VpnHood.AppLib.Settings;
 using VpnHood.Core.Client.Abstractions;
 using VpnHood.Core.Client.VpnServices.Abstractions;
+using VpnHood.Core.Toolkit.Utils;
 
 namespace VpnHood.AppLib.Utils;
 
 public static class StateHelper
 {
-    public static bool IsTcpProxy(AppFeatures features, UserSettings userSettings, ConnectionInfo? connectionInfo, 
+    public static bool IsTcpProxy(AppFeatures features, UserSettings userSettings, ConnectionInfo? connectionInfo,
         bool isReconfiguring)
     {
         var sessionStatus = connectionInfo?.SessionStatus;
@@ -15,7 +16,7 @@ public static class StateHelper
         // let use the current user setting if change is possible
         // it makes sure the UI and the actual setting are in sync during reconfiguration till real state is known
         if (sessionStatus != null && CanChangeTcpProxy(features, sessionInfo))
-            return userSettings.UseTcpProxy; 
+            return userSettings.UseTcpProxy;
 
         return
             sessionStatus?.IsTcpProxy ??
@@ -29,6 +30,17 @@ public static class StateHelper
             return features.IsTcpProxySupported;
 
         return sessionInfo is { IsTcpProxySupported: true, IsTcpPacketSupported: true };
+    }
+
+    public static int? GetProgress(ConnectionInfo? connectionInfo)
+    {
+        var progress = connectionInfo?.ClientStateProgress;
+        if (progress == null ||
+            progress.Value.Total < 3 ||
+            FastDateTime.Now - connectionInfo?.ClientStateChangedTime < TimeSpan.FromSeconds(3))
+            return null;
+
+        return (int)(progress.Value.Done * 100L / progress.Value.Total);
     }
 
 }