+++ Src/Core/VpnHood.Core.Client.Abstractions/ClientState.cs
@@ -6,7 +6,7 @@ public enum ClientState
 {
     None,
     Initializing,
-    FindingServer,
+    FindingReachableServer,
     FindingBestServer,
     Connecting,
     Connected,