+++ Tests/VpnHood.AppLib.Test/Tests/AccessCodeTest.cs
@@ -1,6 +1,7 @@
 ﻿using System.Net.Sockets;
 using System.Text;
 using VpnHood.AppLib.ClientProfiles;
+using VpnHood.Core.Client.Abstractions;
 using VpnHood.Core.Common.Exceptions;
 using VpnHood.Core.Common.Messaging;
 using VpnHood.Core.Common.Tokens;